System Architecture Class Diagram
This is the system architecture class diagram of this project. As shown in the figure, a layered architecture is used to design the input and output, control, model and database by dividing each other. In this way, the stability of the system architecture can be ensured under the MVC architecture, so that the system has a certain tolerance for unknown errors, and the entire system will not collapse due to a module problem.